+++ src/index/khx_reader.hpp
@@ -18,7 +18,7 @@ public:
     uint8_t template_type() const { return template_type_; }
 
     // Check if a k-mer was excluded during index build.
-    bool is_excluded(uint64_t kmer_idx) const {
+    bool is_excluded(uint32_t kmer_idx) const {
         return (bitset_[kmer_idx / 8] >> (kmer_idx % 8)) & 1;
     }
 
@@ -35,7 +35,7 @@ private:
     uint8_t t_ = 0;
     uint8_t template_type_ = 0;
     const uint8_t* bitset_ = nullptr;
-    uint64_t tbl_size_ = 0;
+    uint32_t tbl_size_ = 0;
 };
 
 } // namespace ikafssn
